add document to a safetyMaterial

+++ code/avh.nufm.business.model/src/main/java/avh/nufm/business/model/Invoice.java
@@ -2,6 +2,7 @@ package avh.nufm.business.model;
 
 import java.io.Serializable;
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 
@@ -28,19 +29,19 @@ public class Invoice implements Serializable {
 	private String invoiceTo;
 
 	@Column(name="invoice_total")
-	private double invoiceTotal;
+	private BigDecimal invoiceTotal;
 
-	private double price;
+	private BigDecimal price;
 
-	private double quantity;
+	private BigDecimal quantity;
 
 	@Column(name="sales_tax")
-	private double salesTax;
+	private BigDecimal salesTax;
 
 	private String status;
 
 	@Column(name="sub_total")
-	private double subTotal;
+	private BigDecimal subTotal;
 
 	public Invoice() {
 	}
@@ -85,35 +86,35 @@ public class Invoice implements Serializable {
 		this.invoiceTo = invoiceTo;
 	}
 
-	public double getInvoiceTotal() {
+	public BigDecimal getInvoiceTotal() {
 		return this.invoiceTotal;
 	}
 
-	public void setInvoiceTotal(double invoiceTotal) {
+	public void setInvoiceTotal(BigDecimal invoiceTotal) {
 		this.invoiceTotal = invoiceTotal;
 	}
 
-	public double getPrice() {
+	public BigDecimal getPrice() {
 		return this.price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
 
-	public double getQuantity() {
+	public BigDecimal getQuantity() {
 		return this.quantity;
 	}
 
-	public void setQuantity(double quantity) {
+	public void setQuantity(BigDecimal quantity) {
 		this.quantity = quantity;
 	}
 
-	public double getSalesTax() {
+	public BigDecimal getSalesTax() {
 		return this.salesTax;
 	}
 
-	public void setSalesTax(double salesTax) {
+	public void setSalesTax(BigDecimal salesTax) {
 		this.salesTax = salesTax;
 	}
 
@@ -125,11 +126,11 @@ public class Invoice implements Serializable {
 		this.status = status;
 	}
 
-	public double getSubTotal() {
+	public BigDecimal getSubTotal() {
 		return this.subTotal;
 	}
 
-	public void setSubTotal(double subTotal) {
+	public void setSubTotal(BigDecimal subTotal) {
 		this.subTotal = subTotal;
 	}
